requirements.txt
# Transcription
openai-whisper

# Gemini
google-genai

# Environment
python-dotenv

# Audio (system dependency: ffmpeg must be installed separately)
# Install with: apt install ffmpeg  OR  brew install ffmpeg

# PyTorch with CUDA — install manually based on your CUDA version:
# https://pytorch.org/get-started/locally/
# Example for CUDA 12.1:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
# For CPU only:
#   pip install torch
torch
